+++ src/plafosim/message.py
@@ -164,7 +164,7 @@ class PlatoonAdvertisement(Message):
             destination: int,
             platoon_id: int,
             leader_id: int,
-            platoon_speed: int,
+            platoon_speed: float,
             platoon_lane: int,
             platoon_formation: list,
             platoon_position_front: int,
@@ -193,7 +193,7 @@ class PlatoonAdvertisement(Message):
         return self._leader_id
 
     @property
-    def platoon_speed(self) -> int:
+    def platoon_speed(self) -> float:
         return self._platoon_speed
 
     @property
